alert, input bill, update table products

+++ admin/src/components/UserProfile.jsx
@@ -5,25 +5,50 @@ import { adminProfileData, staffProfileData } from "../data/dummy";
 import Buttonlogout from "@mui/material/Button";
 import axios from "axios";
 import { NavLink, useNavigate } from 'react-router-dom';
+import { Modal } from "antd";
 const UserProfile = ({ isAdmin }) => {
   const user = JSON.parse(localStorage.getItem('user'));
 
 
   const handleLogout = async () => {
     const result = JSON.parse(localStorage.getItem("result"));
-    await axios.post(
-      "http://localhost:4000/users/logout",
-      {
-        refresh_token: result.refresh_token,
+    Modal.confirm({
+      title: 'Xác nhận đăng xuất',
+      content: `Bạn có chắc chắn muốn đăng xuất?`,
+      onOk: async () => {
+        await axios.post(
+          "http://localhost:4000/users/logout",
+          {
+            refresh_token: result.refresh_token,
+          },
+          {
+            headers: {
+              Authorization: `Bearer ${result.access_token}`,
+            },
+          }
+        );
+        localStorage.clear();
+        window.location.reload();
       },
-      {
-        headers: {
-          Authorization: `Bearer ${result.access_token}`,
+      onCancel() {
+        console.log('Cancel');
+      },
+      okButtonProps: {
+        style: {
+          backgroundColor: '#46B5C1',
+          borderColor: '#46B5C1',
         },
-      }
-    );
-    localStorage.clear();
-    window.location.reload();
+      },
+      cancelButtonProps: {
+        style: {
+          backgroundColor: '#FF4D4F',
+          borderColor: '#FF4D4F',
+          color: '#FFFFFF',
+        },
+      },
+      cancelText: 'Đóng',
+      okText: 'Đồng ý',
+    });
   };
   return (
     <div className="nav-item absolute right-1 top-10 bg-white dark:bg-[#42464D] p-5 rounded-lg w-96">
@@ -91,22 +116,22 @@ const UserProfile = ({ isAdmin }) => {
           <div>
             <div>
               {staffProfileData.map((item, index) => (
-                <div
+                <NavLink
                   key={index}
-                  className="flex gap-5 border-b-1 border-color p-4 hover:bg-light-gray cursor-pointer  dark:hover:bg-[#42464D]"
+                  to={`/${item.path}`}
+                  className="flex gap-5 border-b-1 border-color p-4 hover:bg-light-gray cursor-pointer dark:hover:bg-[#42464D]"
                 >
-                  <NavLink
+                  <div
                     style={{ color: item.iconColor, backgroundColor: item.iconBg }}
                     className="text-xl rounded-lg p-3 hover:bg-light-gray"
-                    to={`/${item.path}`}
                   >
                     {item.icon}
-                    <div>
-                      <p className="font-semibold dark:text-gray-200">{item.title}</p>
-                      <p className="text-gray-500 text-sm dark:text-gray-400">{item.desc}</p>
-                    </div>
-                  </NavLink>
-                </div>
+                  </div>
+                  <div>
+                    <p className="font-semibold dark:text-gray-200" style={{marginBottom: '5px'}}>{item.title}</p>
+                    <p className="text-gray-500 text-sm dark:text-gray-400" style={{marginBottom: '5px'}}>{item.desc}</p>
+                  </div>
+                </NavLink>
               ))}
             </div>
             <div className="mt-5">
